treewide: oplus: Fix pointer to integer cast ../../../../../../kernel/realme/mt6785/drivers/misc/mediatek/sensors-1.0/oplus_sensor_devinfo/sensor_devinfo.c:845:39: error: cast to smaller integer type 'uint32_t' (aka 'unsigned int') from 'void *' [-Werror,-Wvoid-pointer-to-int-cast]         DEVINFO_LOG("Ptr2UINT32(p) = %d \n", Ptr2UINT32(p));                                              ^~~~~~~~~~~~~
../../../../../../kernel/realme/mt6785/drivers/misc/mediatek/sensors-1.0/oplus_sensor_devinfo/sensor_devinfo.c:66:25: note: expanded from macro 'Ptr2UINT32'
                        ^~~~~~~~~~~~~

+++ drivers/soc/oplus/system/oplus_project/gki_mtk/oplus_project.c
@@ -22,7 +22,7 @@
 #define SMEM_PROJECT    135
 
 #define UINT2Ptr(n)        (uint32_t *)(n)
-#define Ptr2UINT32(p)    (uint64_t)(p)
+#define Ptr2UINT32(p)    (uintptr_t)(p)
 
 #define PROJECT_VERSION            (0x1)
 #define PCB_VERSION                (0x2)
